backend/requirements.txt
fastapi==0.115.10  # [all]
fastapi-admin
uvicorn
sqlalchemy
pydantic
pyjwt
faiss-cpu
langchain-cli
langchain-openai
langchain-anthropic
langchain-ollama
langchain-core
langchain-community
langserve  # [all]
Pillow
